// src/types/interface/character.tsx
interface Character {
    name: string;
    race: string;
    class: string;
    level: number;
    strength: number;
    dexterity: number;
    constitution: number;
    intelligence: number;
    wisdom: number;
    charisma: number;
    strengthMod: number;
    dexterityMod: number;
    constitutionMod: number;
    intelligenceMod: number;
    wisdomMod: number;
    charismaMod: number;
    armorClass: number;
    initiative: number;
    speed: number;
    maxHP: number;
    spells: string[];
    equipment: string[];
    traits: string[];
    personality: string;
    ideals: string;
    bonds: string;
    flaws: string;
    allies: string;
    backstory: string;
    gold: number;
    age: number;
    height: number;
    weight: number;
    eyes: string;
    skin: string;
    hair: string;
    profBonus: number;
    passiveWisdom: number;
    languages: string[];
    featuresTraits: string[];
    proficiencies: string[];
  }

export default Character;